fix `get_slots` function in utils
tuples are immutable; instead of reassigning on each iter, using
list.append() instead.
Also, the tuple addition just created a really long tuple, instead of a
list of tuples or a tuple of tuples as expected.

datetime.strptime() gives a datetime object, so it renders as
'1900-01-01 xx:xx'; using time() to strip away the date.

Both values of the tuple are used for rendering the form, so each tuple
should be of the form (value, displayed_option). Putting the same value
at the first position (0) in all tuples makes all the options identical.
Using the start_time.time() as the value.

Thus, the end-times are not stored, but since the duration is known that
can be easily extrapolated.